relocate utils to utilities folder
